Merge `8.1.x` in `develop`

+++ tests/UI/campaigns/productV2/functional/04_CRUDPackOfProducts.ts
@@ -27,6 +27,7 @@ import {
 import ProductData from '@data/faker/product';
 import Employees from '@data/demo/employees';
 import Products from '@data/demo/products';
+import {ProductPackOptions} from '@data/types/product';
 
 const baseContext: string = 'productV2_functional_CRUDPackOfProducts';
 
@@ -53,18 +54,18 @@ describe('BO - Catalog - Products : CRUD pack of products', async () => {
     status: false,
   });
 
-  const editPackData: object = {
+  const editPackData: ProductPackOptions = {
     quantity: 100,
     minimalQuantity: 2,
     packQuantitiesOption: 'Decrement pack only',
   };
 
   // Data to edit the product price
-  const pricingData: object = {
-    price: 15,
+  const pricingData: ProductData = new ProductData({
+    priceTaxExcluded: 15,
     taxRule: 'FR Taux standard (20%)',
-    priceTaxIncl: 18,
-  };
+    price: 18,
+  });
 
   const editProductData: ProductData = new ProductData({
     type: 'pack',
@@ -88,14 +89,22 @@ describe('BO - Catalog - Products : CRUD pack of products', async () => {
   before(async function () {
     browserContext = await helper.createBrowserContext(this.browser);
     page = await helper.newTab(browserContext);
-    await files.generateImage(newProductData.coverImage);
-    await files.generateImage(newProductData.thumbImage);
+    if (newProductData.coverImage) {
+      await files.generateImage(newProductData.coverImage);
+    }
+    if (newProductData.thumbImage) {
+      await files.generateImage(newProductData.thumbImage);
+    }
   });
 
   after(async () => {
     await helper.closeBrowserContext(browserContext);
-    await files.deleteFile(newProductData.coverImage);
-    await files.deleteFile(newProductData.thumbImage);
+    if (newProductData.coverImage) {
+      await files.deleteFile(newProductData.coverImage);
+    }
+    if (newProductData.thumbImage) {
+      await files.deleteFile(newProductData.thumbImage);
+    }
   });
 
   // 1 - Create product
@@ -245,7 +254,7 @@ describe('BO - Catalog - Products : CRUD pack of products', async () => {
           it('should choose the searched product', async function () {
             await testContext.addContextItem(this, 'testIdentifier', `chooseProduct${index}`, baseContext);
 
-            const isListOfProductVisible = await packTab.selectProductFromList(page);
+            const isListOfProductVisible = await packTab.selectProductFromList(page, 1);
             await expect(isListOfProductVisible).to.be.true;
           });
         }
@@ -265,7 +274,7 @@ describe('BO - Catalog - Products : CRUD pack of products', async () => {
             await expect(result.image).to.contains(test.args.product.defaultImage),
             await expect(result.name).to.equal(test.args.productToChooseName),
             await expect(result.reference).to.equal(`Ref: ${test.args.product.reference}`),
-            await expect(parseInt(result.quantity, 10)).to.equal(1),
+            await expect(result.quantity).to.equal(1),
           ]);
         });
       });
